apps/research-ui/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.2.0
pydantic[email]>=2.0.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<5
httpx>=0.27.0
pytest>=8.0.0
pytest-anyio>=0.0.0
anyio[trio]>=4.0.0
